+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Option to select IP Family to use for LINSTOR control traffic.
+
 ## [v2.6.0] - 2024-09-04
 
 ### Added